staticfiles operations
