+++ fab/fabfile.py
@@ -58,7 +58,7 @@ from const import (
     RELEASE_RECORD,
     RSYNC_EXCLUDE,
 )
-from operations import db
+from operations import db, staticfiles
 
 
 if env.ssh_config_path and os.path.isfile(os.path.expanduser(env.ssh_config_path)):
@@ -573,11 +573,11 @@ def _deploy_without_asking():
         _execute_with_timing(db.ensure_preindex_completion)
 
         # handle static files
-        _execute_with_timing(version_static)
-        _execute_with_timing(_bower_install)
-        _execute_with_timing(_npm_install)
-        _execute_with_timing(_do_collectstatic)
-        _execute_with_timing(_do_compress)
+        _execute_with_timing(staticfiles.version_static)
+        _execute_with_timing(staticfiles.bower_install)
+        _execute_with_timing(staticfiles.npm_install)
+        _execute_with_timing(staticfiles.collectstatic)
+        _execute_with_timing(staticfiles.compress)
 
         _set_supervisor_config()
 
@@ -589,12 +589,12 @@ def _deploy_without_asking():
             _execute_with_timing(stop_celery_tasks)
         _execute_with_timing(_migrate)
 
-        _execute_with_timing(do_update_translations)
+        _execute_with_timing(staticfiles.update_translations)
         _execute_with_timing(db.flip_es_aliases)
 
         # hard update of manifest.json since we're about to force restart
         # all services
-        _execute_with_timing(update_manifest)
+        _execute_with_timing(staticfiles.update_manifest)
         _execute_with_timing(clean_releases)
     except PreindexNotFinished:
         mail_admins(
@@ -836,9 +836,9 @@ def clean_releases(keep=3):
 @task
 def force_update_static():
     _require_target()
-    execute(_do_collectstatic, use_current_release=True)
-    execute(_do_compress, use_current_release=True)
-    execute(update_manifest, use_current_release=True)
+    execute(staticfiles.collectstatic, use_current_release=True)
+    execute(staticfiles.compress, use_current_release=True)
+    execute(staticfiles.update_manifest, use_current_release=True)
     silent_services_restart(use_current_release=True)
 
 
@@ -1014,91 +1014,6 @@ def _migrations_exist():
             # failed to return a value python could parse into an int
             return True
         return n_migrations > 0
-
-
-@parallel
-@roles(ROLES_STATIC)
-def _do_compress(use_current_release=False):
-    """Run Django Compressor after a code update"""
-    venv = env.virtualenv_root if not use_current_release else env.virtualenv_current
-    with cd(env.code_root if not use_current_release else env.code_current):
-        sudo('{}/bin/python manage.py compress --force -v 0'.format(venv))
-        sudo('{}/bin/python manage.py purge_compressed_files'.format(venv))
-    update_manifest(save=True, use_current_release=use_current_release)
-
-
-@parallel
-@roles(ROLES_STATIC)
-def _do_collectstatic(use_current_release=False):
-    """Collect static after a code update"""
-    venv = env.virtualenv_root if not use_current_release else env.virtualenv_current
-    with cd(env.code_root if not use_current_release else env.code_current):
-        sudo('{}/bin/python manage.py collectstatic --noinput -v 0'.format(venv))
-        sudo('{}/bin/python manage.py fix_less_imports_collectstatic'.format(venv))
-        sudo('{}/bin/python manage.py compilejsi18n'.format(venv))
-
-
-@parallel
-@roles(ROLES_STATIC)
-def _bower_install(use_current_release=False):
-    with cd(env.code_root if not use_current_release else env.code_current):
-        sudo('bower prune --production --config.interactive=false')
-        sudo('bower update --production --config.interactive=false')
-
-
-@parallel
-@roles(ROLES_DJANGO)
-def _npm_install():
-    with cd(env.code_root):
-        sudo('npm prune --production')
-        sudo('npm install --production')
-        sudo('npm update --production')
-
-
-@roles(ROLES_DJANGO)
-@parallel
-def update_manifest(save=False, soft=False, use_current_release=False):
-    """
-    Puts the manifest.json file with the references to the compressed files
-    from the proxy machines to the web workers. This must be done on the WEB WORKER, since it
-    governs the actual static reference.
-
-    save=True saves the manifest.json file to redis, otherwise it grabs the
-    manifest.json file from redis and inserts it into the staticfiles dir.
-    """
-    withpath = env.code_root if not use_current_release else env.code_current
-    venv = env.virtualenv_root if not use_current_release else env.virtualenv_current
-
-    args = ''
-    if save:
-        args = ' save'
-    if soft:
-        args = ' soft'
-    cmd = 'update_manifest%s' % args
-    with cd(withpath):
-        sudo('{venv}/bin/python manage.py {cmd}'.format(venv=venv, cmd=cmd),
-            user=env.sudo_user
-        )
-
-
-@roles(set(ROLES_STATIC + ROLES_DJANGO))
-@parallel
-def version_static():
-    """
-    Put refs on all static references to prevent stale browser cache hits when things change.
-    This needs to be run on the WEB WORKER since the web worker governs the actual static
-    reference.
-
-    """
-
-    cmd = 'resource_static'
-    with cd(env.code_root):
-        sudo(
-            'rm -f tmp.sh resource_versions.py; {venv}/bin/python manage.py {cmd}'.format(
-                venv=env.virtualenv_root, cmd=cmd
-            ),
-            user=env.sudo_user
-        )
 
 
 def _rebuild_supervisor_conf_file(conf_command, filename, params=None):
@@ -1281,20 +1196,6 @@ def stop_celery_tasks():
     _require_target()
     with cd(env.code_root):
         sudo('scripts/supervisor-group-ctl stop celery')
-
-
-@roles(ROLES_ALL_SRC)
-@parallel
-def do_update_translations():
-    with cd(env.code_root):
-        update_locale_command = '{virtualenv_root}/bin/python manage.py update_django_locales'.format(
-            virtualenv_root=env.virtualenv_root,
-        )
-        update_translations_command = '{virtualenv_root}/bin/python manage.py compilemessages'.format(
-            virtualenv_root=env.virtualenv_root,
-        )
-        sudo(update_locale_command)
-        sudo(update_translations_command)
 
 
 @task
